fixed typos and style in nyc_taxi_fares and eletricity projects
